+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 
 - Add client specific `map.client.` and `tmp.client.` variable namespaces
+- Add game addon remote client compatibility check (to check if client has the DLC if the server has it)
 
 ### Fixed